% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellipse_grid.R
\name{ellipse_grid}
\alias{ellipse_grid}
\title{Ellipse grid}
\usage{
ellipse_grid(N, radius, center = c(0, 0))
}
\arguments{
\item{N}{The number of subintervals to divide the smaller radius into.}

\item{radius}{A vector of length two giving the radius of each length.}

\item{center}{A vector of length two giving the center of the ellipse.}
}
\value{
Returns a matrix with the x and y positions of each point in the grid.
}
\description{
Create a grid of points within an ellipse.
}
\details{
Code taken from \url{https://people.sc.fsu.edu/~jburkardt/cpp_src/ellipse_grid/ellipse_grid.html}.
}
\examples{
p = shapegrid::ellipse_grid(25, c(1,1), c(0,0))
plot(p)
}
