% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triangle_grid.R
\name{triangle_grid}
\alias{triangle_grid}
\title{Triangle grid}
\usage{
triangle_grid(N, vertices)
}
\arguments{
\item{N}{The number of subintervals each side should be divided into i.e. there are (N+1) points on a side.}

\item{vertices}{A vector of length six giving the (x,y) coordinates of each vertex of the triangle or a matrix with two rows and three columns of vertex coordinates.}
}
\value{
Returns a matrix with the x and y positions of each point in the grid.
}
\description{
Create a grid of points within a triangle
}
\details{
Code taken from \url{https://people.sc.fsu.edu/~jburkardt/cpp_src/triangle_grid/triangle_grid.html}.
}
\examples{
v = cbind(c(0,0), c(1,1), c(2,0))
p = shapegrid::triangle_grid(10, v)
plot(p)
}
